Initial commit: added source files for calculator
This is my first project. I was trying to make a calculator, but the problem was I didn't know how to parse an expression until I learned about "shunting yard algorithm". And after that, everything became easier. I've tested it thoroughly, but it might still contain some bugs.......If the program crashes just press Alt + F4.   :)